Add schemes

+++ src/models/transaction-features/quest-dispute/QuestDisputeDecisionData.ts
@@ -27,7 +27,7 @@ export class QuestDisputeDecisionData extends Model {
 
   @Column(DataType.STRING) decision: QuestDisputeDecision;
 
-  @Column(DataType.INTEGER) gasPriceAtMoment: string;
+  @Column(DataType.BIGINT) gasPriceAtMoment: string;
   @Column({ type: DataType.INTEGER, allowNull: false }) status: TransactionStatus;
 
   @Column(DataType.STRING) error: string;
